upgrade JSON-FG to v0.1
* reflect changes (when -> time, where -> place)
* add support for GeoJSON compatibility mode, hide format in HTML
* add support for polyhedron geometries
* rename ogcapi-features-json-fg to ogcapi-features-jsonfg to improve package names
