--- ogcapi-draft/ogcapi-features-json-fg/src/main/java/de/ii/ogcapi/features/json/fg/app/JacksonSubTypeIdsJsonFg.java
@@ -1,36 +1,0 @@
-/*
- * Copyright 2022 interactive instruments GmbH
- *
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- */
-package de.ii.ogcapi.features.json.fg.app;
-
-import com.github.azahnen.dagger.annotations.AutoBind;
-import com.google.common.collect.ImmutableList;
-import de.ii.ogcapi.features.json.fg.domain.JsonFgConfiguration;
-import de.ii.ogcapi.foundation.domain.ExtensionConfiguration;
-import de.ii.xtraplatform.base.domain.ImmutableJacksonSubType;
-import de.ii.xtraplatform.base.domain.JacksonSubTypeIds;
-import java.util.List;
-import javax.inject.Inject;
-import javax.inject.Singleton;
-
-@Singleton
-@AutoBind
-public class JacksonSubTypeIdsJsonFg implements JacksonSubTypeIds {
-
-  @Inject
-  public JacksonSubTypeIdsJsonFg() {}
-
-  @Override
-  public List<JacksonSubType> getSubTypes() {
-    return ImmutableList.of(
-        ImmutableJacksonSubType.builder()
-            .superType(ExtensionConfiguration.class)
-            .subType(JsonFgConfiguration.class)
-            .id(ExtensionConfiguration.getBuildingBlockIdentifier(JsonFgConfiguration.class))
-            .build());
-  }
-}
