+++ ogcapi-draft/ogcapi-features-jsonfg/build.gradle
