+++ ogcapi-draft/ogcapi-features-jsonfg/src/main/java/de/ii/ogcapi/features/jsonfg/domain/package-info.java
@@ -1,0 +1,10 @@
+/**
+ * Copyright 2022 interactive instruments GmbH
+ *
+ * <p>This Source Code Form is subject to the terms of the Mozilla Public License, v. 2.0. If a copy
+ * of the MPL was not distributed with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+@AutoModule(single = true, encapsulate = true)
+package de.ii.ogcapi.features.jsonfg.domain;
+
+import com.github.azahnen.dagger.annotations.AutoModule;
