+++ ogcapi-stable/ogcapi-features-html/src/main/java/de/ii/ogcapi/features/html/app/FeaturesFormatHtml.java
@@ -544,7 +544,9 @@ public class FeaturesFormatHtml
 
     featureTypeDataset.formats =
         links.stream()
-            .filter(link -> Objects.equals(link.getRel(), "alternate"))
+            .filter(
+                link ->
+                    Objects.equals(link.getRel(), "alternate") && !link.getTypeLabel().isBlank())
             .sorted(Comparator.comparing(link -> link.getTypeLabel().toUpperCase()))
             .map(link -> new NavigationDTO(link.getTypeLabel(), link.getHref()))
             .collect(Collectors.toList());
@@ -589,7 +591,9 @@ public class FeaturesFormatHtml
 
     featureCollectionView.formats =
         links.stream()
-            .filter(link -> Objects.equals(link.getRel(), "alternate"))
+            .filter(
+                link ->
+                    Objects.equals(link.getRel(), "alternate") && !link.getTypeLabel().isBlank())
             .sorted(Comparator.comparing(link -> link.getTypeLabel().toUpperCase()))
             .map(link -> new NavigationDTO(link.getTypeLabel(), link.getHref()))
             .collect(Collectors.toList());
