+++ ogcapi-stable/ogcapi-html/src/main/java/de/ii/ogcapi/html/domain/OgcApiView.java
@@ -53,7 +53,8 @@ public abstract class OgcApiView extends View {
 
   public List<NavigationDTO> getFormats() {
     return links.stream()
-        .filter(link -> Objects.equals(link.getRel(), "alternate"))
+        .filter(
+            link -> Objects.equals(link.getRel(), "alternate") && !link.getTypeLabel().isBlank())
         .sorted(Comparator.comparing(link -> link.getTypeLabel().toUpperCase()))
         .map(link -> new NavigationDTO(link.getTypeLabel(), link.getHref()))
         .collect(Collectors.toList());
